Add fields to the request class.
These fields will be used to store the properties that will be
needed to process a given request. All fields are private at this
point, and accessors will be added at some point. With one
possible exception ($params), mutators should not be needed as the
request object will be treated as an immutable data store.