Calculate and initialize $basePath property in Request
Forgotten in previous commit. The $basePath property is calculated
by taking the directory name of the URL path to index.php (which
will be processing all requests). Here, $basePath is where you have
to navigate to, from the document root, to get to the location of
index.php. This will be used to remove the base path from REQUEST_URI
strings, so that we don't have to worry about where our project is
located, in relation to the document root.