Add simple accessors for all remaining fields in the Request class
These accessors build in safety for values which may not be set.
This is done through use of the $_ARR['key'] ?? 'SomeDefault'
pattern which will take the value of $_ARR['key'] if it is set,
or will substitute 'SomeDefault' if it is not. Note: this avoids
the need for isset() checks on these value, but not for empty()
checks. The default apploed for all values here is NULL.

The above pattern does not negate the need for validation of user
values, it simply removes worry that a PHP warnign will be generated
if you attempt to access, as example, a key in an associative array
which does not exist.