Add raw send() method for a Response
This method will take care of setting the current status, any defined
headers, and of sending the block of provided content. There is an
argument for making this method private, to prvent is being called by
outside code, but calling from outside is not 'dangerous' and may be
needed in some cases, so we will leave it public.