Fix bug in Response redirect() method.
The redirect method was using code from an older commit, which
has since been refactored. The method was calling a non-existant
mehtod of the router stored in a response, to get the request base
URL. This now exists in the request object, to the internal router
reference in the response has now been replaced with the request.