Improve how headers are parsed by adding 'lazy' parsing
Previously, headers were always parsed and stored in the Request
constructor. Now, they are only parsed and stored on the first
call of the header() getter -- this way, if no headers are needed,
the script will not have to go through the process of parsing them.
If they are needed, the first call to the header() getter will
initiate parsing.