Add utility ConfigFile::getContent() to load config files in a safe way
When used, this will load a config file located at [project]/config.php
and will cache the contents so that additional calls will not result in
additional loads from disk.

For this to work, a config file must be located at [project]/config.php
and must RETURN its content. Each call to ConfigFile::getContent()
will return the contents of the config file (from a cache on all call
after the initial call).